Backend/flask/requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
filelock==3.10.7
Flask==2.2.3
Flask-Cors==3.0.10
future==0.18.3
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
protobuf==3.20.3
PySocks==1.7.1
python-crontab==2.7.1
python-dateutil==2.8.2
requests==2.28.2
requests-file==1.5.1
six==1.16.0
tldextract==3.4.0
urllib3==1.26.15
Werkzeug==2.2.3
zope.event==4.6
zope.interface==6.0
